Refactor code for readability
