+++ src/style.scss
@@ -69,14 +69,14 @@ body {
         user-select: none;
         cursor: pointer;
         z-index: 1;
-        //box-shadow: 0 0 1em rgba(white, 0.2);
+        box-shadow: 0 0 1em rgba(0, 0, 0, 0.2);
         transition:
           transform 0.1s cubic-bezier(0.5, 0, 0.5, 1),
           box-shadow 0.2s;
 
-        //&:hover {
-        //  box-shadow: 0 0 2em rgba(white, 0.3);
-        //}
+        &:hover {
+          box-shadow: 0 0 2em rgba(0, 0, 0, 0.3);
+        }
 
         outline: none;
